UI reworked and optimized. Fixed bugs after playtest. Main menu added, Mouse sensitivity settings added

+++ Assets/Scripts/Core_Utils.cs
@@ -34,7 +34,7 @@ public static class Core_Utils
     {
         return Vector3.Distance(current.position + new Vector3(0, 1, 0), Player.transform.position) < 2;
     }
-    public static bool isCloseToPlayer(Transform current, int distance)
+    public static bool isCloseToPlayer(Transform current, float distance)
     {
         return Vector3.Distance(current.position + new Vector3(0, 1, 0), Player.transform.position) < distance;
     }
